Added new and deactivate functions for AssetManagers. AMAAS-385
